+++ src/admin-template-configure-columns/requisition-template-admin.controller.js
@@ -90,7 +90,7 @@
          * Redirects user to template list view page.
          */
         function goToTemplateList() {
-            $state.go('openlmis.administration.programs');
+            $state.go('openlmis.administration.templates');
         }
 
         /**