+++ src/admin-template-list/admin-template-list.controller.spec.js
@@ -15,7 +15,8 @@
 
 describe('TemplateListAdminController', function() {
 
-    var vm, template, program, ProgramDataBuilder, $controller;
+    var vm, template, program, ProgramDataBuilder, TemplateDataBuilder, $controller,
+        FacilityTypeDataBuilder;
 
     beforeEach(function() {
         module('admin-template-list');
@@ -23,37 +24,17 @@ describe('TemplateListAdminController', function() {
         inject(function($injector) {
             $controller = $injector.get('$controller');
             ProgramDataBuilder = $injector.get('ProgramDataBuilder');
+            TemplateDataBuilder = $injector.get('TemplateDataBuilder');
+            FacilityTypeDataBuilder = $injector.get('FacilityTypeDataBuilder');
         });
 
-        program = new ProgramDataBuilder().build();
+        program = new ProgramDataBuilder().withId('program-1').build();
 
-        template = jasmine.createSpyObj('template', ['$save']);
-        template.id = '1';
-        template.programId = program.id;
-        template.columnsMap = {
-            remarks: {
-                displayOrder: 1,
-                isDisplayed: true,
-                label: 'Remarks'
-            },
-            total: {
-                displayOrder: 2,
-                isDisplayed: true,
-                label: 'Total'
-            },
-            stockOnHand: {
-                displayOrder: 3,
-                isDisplayed: true,
-                label: "Stock on Hand"
-            },
-            averageConsumption: {
-                name: 'averageConsumption',
-                displayOrder: 4,
-                isDisplayed: true,
-                label: "Average Consumption"
-            }
-        };
-        template.numberOfPeriodsToAverage = 3;
+        template = new TemplateDataBuilder()
+            .withFacilityTypes([
+                new FacilityTypeDataBuilder().build(),
+                new FacilityTypeDataBuilder().buildDistrictHospital().build()
+            ]).build();
 
         vm = $controller('TemplateListAdminController', {
             programs: [program],
@@ -67,6 +48,11 @@ describe('TemplateListAdminController', function() {
             vm.$onInit();
             expect(vm.programs).toEqual([program]);
             expect(vm.templates).toEqual([template]);
+        });
+
+        it('should add templates to program', function() {
+            vm.$onInit();
+            expect(vm.programs[0].templates).toEqual([template]);
         });
     });